Checker

Description

Documentation

checkMove :: Board -> [Char] -> [Char] -> Piece -> Piece -> PColor -> Bool

checkMove

Takes a board, starting field coordinates, target field coordinates, current starting field Piece, current target field Piece and current players' color.

Returns False when:

color of starting Piece is different from players' (player can't move opponents' piece)

color of starting Piece is the same as target Piece (player can't capture own piece)

starting coordinates are the same as target coordinates (player can't stand in a place in a loop)

type of the staring Piece is other then available (Piece is non existing)

| Returns True when Piece types' legality checkers returns True.

kingChecker :: Board -> [Char] -> [Char] -> Bool

queenChecker :: Board -> [Char] -> [Char] -> Bool

rookChecker :: Board -> [Char] -> [Char] -> Bool

bishopChecker :: Board -> [Char] -> [Char] -> Bool

knightChecker :: Board -> [Char] -> [Char] -> Bool

pawnChecker :: Board -> [Char] -> [Char] -> PColor -> Bool

emptyPath :: [Field] -> [[Char]] -> [[Char]] -> [[Char]]

emptyPath

Out of a list of legal moves' coordinates, creates a path of unoccupied coordines, starting from the starting coordinate to a first encountered occupied coordinate. Adds the coordinate to the list, as it can be occupied by a Piece to capture. Used by Queen, Rook and Bishop Piece type.

>>> emptyPath ["6b", "6c", "6d", "6e", "6f"]
["6b", "6c", "6d"]

emptyField :: [Field] -> [[Char]] -> [[Char]]

emptyField

Out of a list of moves' coordinates, checks if coordinates can exist as legal, by being unoccupied.

Used by Pawn Piece type.

opponentsPiece :: [Field] -> [[Char]] -> PColor -> [[Char]]

opponentsPiece

Out of a list of moves' coordinates, checks if coordinates can exist as legal, by being occupied by an opponents' piece.

Used by Pawn Piece type.